Report timers/histograms, add hyphen prefix support
In existing postmates systems there is a need to parse metrics with
names like

    source.machine-metric.name

as being a metric named 'metric.name' with source 'source.machine'.
The etsy/statsd backend has been modified to take a regex to fiddle
with this but I've hard coded it in the expectation that future
backends will be programmable by the end-user.

This commit:

  * Fixes #4
  * Fixes #7

Signed-off-by: Brian L. Troutwine <blt@postmates.com>