Transition hand-rolled statsd parser to LALRPOP
A hand-rolled parser is a cute trick but isn't very maintainable
over the long-term. The previous implementation also wasn't very
efficient owing to a fair amount of allocations happening and
continual seeking.

This commit introduces LALRPOP, a LR(1) parser generator. The
grammer is defined in src/metrics/statsd.lalrpop and build will
generate the rust code automatically.

This commit is related to #6 but does not resolve it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>