Transition hand-rolled statsd parser to LALRPOP (#13)
* Transition hand-rolled statsd parser to LALRPOP

A hand-rolled parser is a cute trick but isn't very maintainable
over the long-term. The previous implementation also wasn't very
efficient owing to a fair amount of allocations happening and
continual seeking.

This commit introduces LALRPOP, a LR(1) parser generator. The
grammer is defined in src/metrics/statsd.lalrpop and build will
generate the rust code automatically.

This commit is related to #6 but does not resolve it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add '_' as a valid member of stat name

This change comes via request by John. "We should take this
opportunity to allow _ here as well. Folks have been desiring
this "feature" since day 1 of statsd."

Signed-off-by: Brian L. Troutwine <blt@postmates.com>