Cernan will now accept arbitrary metadata tags on CLI (#18)
This update allows the user to specify a comma separated list of
metadata tags on cernan startup. This supplants --metric_source in
favor of tags. Old cernan default behaviour can be achieved by
setting

    --tags source=cernan

If you wish to have the above and include a metadata tag foo=bar this
would look like

    --tags source=cernan,foo=bar
    --tags foo=bar,source=cernan

Signed-off-by: Brian L. Troutwine <blt@postmates.com>