Begins the process of normalizing backends
This commit pushes the 'backends' notion into each appropriate
backend with the long-term ambition of allowing 'backends' to be
user-implemented in lua or similar. This does result in some
memory duplication, though I think I will use Arc for Metrics,
where the duplicate occurs presently.

This differntiation of backends allows the wavefront backend to
push exact points, as well as aggregates, allowing the full
utility of wavefront's timeseries language to come to bear.

Two things to improve:

  * flushing and parsing happen in the same thread, a
    potential bottleneck
  * Metrics are duplicated between backends when they likely
    do not need to be

bad_messages / total_messages have been removed for want of a
global place for their tally.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>