Significantly reduce allocations
This commit makes two changes that reduce total allocations. It
uses Rc<> to surround Metrics so that these do not need to be
cloned when they enter a backend. Also, we are now making use of
servo's string_cache::Atom which forces an allocation only the
first time a string is seen. Our standard use-case will be for
many metrics of the same name to be reported over and over.

This resolves issue #6.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>