Integrate graphite into cernan
This commit introduces a graphite interface for cernan. The points
are distributed as expected to existing backends. Primary
modification is the introduce of MetricKind::Raw, of which all
graphite metrics are. This MetricKind will only ever be
aggregated as a guage or, in the case of supporting backends,
be queued for exact delivery.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>