Move from docopts to clap
docopts is really lovely but it's limited in terms of some user
ergonomics. With clap it's easier to build something better for
the user but the internal datastructure we have to pass around
is hard to build.

I've nuked some tests as a result. I feel kind of bad about that.
Abstracting this would be possible but passing the exact structure
lets us avoid some allocations on startup.

Dunno. Feel of two minds about nuking those tests.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>