Graphite (#30)
* Integrate graphite into cernan

This commit introduces a graphite interface for cernan. The points
are distributed as expected to existing backends. Primary
modification is the introduce of MetricKind::Raw, of which all
graphite metrics are. This MetricKind will only ever be
aggregated as a guage or, in the case of supporting backends,
be queued for exact delivery.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* cernan learns how to log

This commit introduces a new -v flag to the cernan cli. Multiple
-vv will kick the logging level of cernan up higher and higher,
all the way into 'trace' mode. This is intended to make verification
of cernan more convenient and allow end-users to debug their systems.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add more trace/debug info, buffered TCP

This commit moves the reading from the graphite TCP socket to use
buffered reading. Per-line messages are put into the channel, a
process which is non-optimal in terms of allocations.

More trace/debug statements have been added to cernan as well.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>