Push message parsing into source threads
Cernan needs to evolve along the lines of the main thread
simply being a very fast router. This becomes especially
pressing when routing from arbitrary source to arbitrary
sink will be available to the end-user. (See #39.)

This commit does not remove the burden of flushing each
backend from the main thread but it does remove the
burden of parsing _and_ flushing. Rc sync has been
changed to Arc to allow a vector of metrics to be
sendable.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>