All backends now run in separate threads
This work is a follow on from the last commit and moves us closer
to the world in which the main thread is a very fast router. There
is little routing going on here as you can see because we have not
yet fully addressed issue #39. We're on a good foot to do it though
that is for sure.

On that day, many Shubs and Zuuls will know what it is to be roasted
in the depths of a Sloar that day, I can tell you!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>