Allow wavefront backend to be specified as DNS name
This commit pulls in the dns-lookup library--which exports the
unstable API present in the stdlib and marks it as stable--to
allow the wavefront proxy host to be specified as a DNS name.

The DNS -> IP lookup is done only at startup and is cached for
the lifetime of the cernan instance.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>