Allow configuration to be specified via CLI
This fixes a backwards compatability issue with cernan's CLI
interface. In particular, with the introduction of the toml
configuration file reading I was too eager to be rid of the
CLI switches and removed them. This makes the upgrade path
for cernan more difficult in the 0.2.x series, requiring
patches to be targeted directly for stable, rather than
master.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>