Enable DNS lookups for wavefront backend (#54)
* Update dependencies ahead of inclusion of dns library

The dependencies of cernan need to be bumped in some cases
to allow for the inclusion of the dns-lookup library. In
particular, the libc dep held by clap is distinct from that
needed by up-to-date dns-lookup. Went ahead and upgraded
other libraries at the suggestion of 'cargo outdated -R'

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Allow wavefront backend to be specified as DNS name

This commit pulls in the dns-lookup library--which exports the
unstable API present in the stdlib and marks it as stable--to
allow the wavefront proxy host to be specified as a DNS name.

The DNS -> IP lookup is done only at startup and is cached for
the lifetime of the cernan instance.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>