Allow configuration to be specified via CLI (#56)
* Allow configuration to be specified via CLI

This fixes a backwards compatability issue with cernan's CLI
interface. In particular, with the introduction of the toml
configuration file reading I was too eager to be rid of the
CLI switches and removed them. This makes the upgrade path
for cernan more difficult in the 0.2.x series, requiring
patches to be targeted directly for stable, rather than
master.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Pump out more helpful conversion error messages

As pointed out by John, the existing use of `unwrap()` in the config
module makes it difficult for users to know what they're doing wrong.
I've flipped `unwrap()` over to `expect()` and given the user some
kind of hint.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>