Include count outputs with histograms / timers
Previously we relied on the wavefront backend to accept all points
and that this would allow us to get percentile information in
addition to counts. That's not true. A previous commit put us back
on a good foot with regard to counts and this puts us in a good
place with regard to counts.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>