Introduce quality of service per metric type
This commit introduces two distinction notions into cernan, that
of per second snapshots of key aggregates and granular reporting
of metrics, called quality of service per John's nomenclature.

Every one second we emit a Snapshot even to sinks. In the case of
the wavefront sink, this means that it will, when appropriate,
take the point-in-time value of its gauges and counters, store
them and clear these values. Summaries--timers and histograms--
are not subject to clearing. The quality of service determines
how often we snapshot.

By default, QOS for a counter is 1 second. That means, if we have
on point enter cernan for a counter per second, there will be one
snapshot per second. QOS for a histogram is 60 seconds, meaning
there will be one snapshot per minute.

This commit removes the librato sink.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>