Move away from mpsc::channel to durable mpmc
This is a largish bit of work to introduce a new mpmc_log queue,
modeled on the API that mpsc::channel provides. The underlying
structure is an append-only file that is written to by a Sender
and read from by a Receiver. The structure of the file is
simplistic. Blobs are offset by a u32 value which gives the length
of the blob:

   u32  | payload_size_in_bytes
   [u8] | payload of size payload_size_in_bytes
   u32  | payload_size_in_bytes
   [u8] | payload of size payload_size_in_bytes
   ...

Right at this very minute the underlying file will grow without
bound. mpmc_log is not generic--though it could be--and is written
with the intention of requiring only serde serializable data. We use
servo's bincode to serialize/deserialize blobs.

This resolves #90. Cernan now allocates 1MB per Receiver and that's
pretty well it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>