Normalize counter snapshot to 1 / sec.
This commit adjusts the way counter snapshots work. Previously
we would sum in a bucket one counter QOS wide and report this out.
What we _wanted_ to do was sum in a QOS wide bucket and report
out a value normalized to 1 / sec. We do this now by dividing by
the bucket width.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>