Bump cernan to 3.0.1
This release corrects the way we report counts with a non-one QOS
and incidentally includes VERY VERY experimental support for reading
from files. DO NOT USE THIS. The downside of this file-reading
support is that we will, for very fast files, buffer too much in
memory and swamp your box.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>