Remove reliance on created metadata
As implemented in the standard library the created metadata is not
available on Linux systems. That's okay though. Since we're naming
these files in order we only need to compare by the path name.

With the new project organization, doc tests are running reliably
and this commit repairs them as well.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>