Durable MPSC (#96)
* Move away from mpsc::channel to durable mpmc

This is a largish bit of work to introduce a new mpmc_log queue,
modeled on the API that mpsc::channel provides. The underlying
structure is an append-only file that is written to by a Sender
and read from by a Receiver. The structure of the file is
simplistic. Blobs are offset by a u32 value which gives the length
of the blob:

   u32  | payload_size_in_bytes
   [u8] | payload of size payload_size_in_bytes
   u32  | payload_size_in_bytes
   [u8] | payload of size payload_size_in_bytes
   ...

Right at this very minute the underlying file will grow without
bound. mpmc_log is not generic--though it could be--and is written
with the intention of requiring only serde serializable data. We use
servo's bincode to serialize/deserialize blobs.

This resolves #90. Cernan now allocates 1MB per Receiver and that's
pretty well it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Rename mpmc_log to mpmc

The new name doesn't really give hint that mpmc_log is a durable
store but it's sure easier to type.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Extract byteoder crate

With our serialization being done by bincode we only need to
handle reading an writing a big-endian u32 to disk. This does
not require an external crate.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove commented out code

I unintentionally left this in place and John pointed out that it
could be removed. It's now removed!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* No longer consume disk space infinitely

This commit moves the mpsc duable channel to use a finite amount
of disk space. This is done by breaking the former single file
append-only log into files of relatively bounded size without
overt synchronization between sender threads. The receiver thread
has gone solo, meaning we are free to easily delete a fully read
through log file.

The implementation in mpsc is effectively a state machine with many
retries.

This resolves #92.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce quickcheck, benchmarking

This commit introduces a limited amount of quickcheck testing in
the mpsc module, benchmarking for same. The source files have been
re-arranged to allow for this but there are no changes to building.
Of interest, 'cargo bench' will now do a sensible thing.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove reliance on created metadata

As implemented in the standard library the created metadata is not
available on Linux systems. That's okay though. Since we're naming
these files in order we only need to compare by the path name.

With the new project organization, doc tests are running reliably
and this commit repairs them as well.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* cargo fmt run

This commit is purely formatting oriented. cargo fmt is a neat
tool and maybe we should have a CI system take care of this.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Clippy suggestions

This commit incorporates the suggestions of the clippy linting
tool. The code that gets generated in some cases is slightly
more efficient.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>