No more extra snapshot thread
All metrics carry in themselves a timestamp. With this, it's
possible for the aggregation module--buckets--to be aware of
if a metric is inside or outside the bounds of the metric's
QOS. Because this implementation was complicated we skipped
it previously and relied on a snapshotting thread--which
fired on one second intervals--to accomplish kinda sorta the
same thing.

This commit removes the snapshotting thread, the snapshot
event and pushes the concept of snapshotting into the metric
aggregtor directly. What we don't do is obey QOS notions yet
which will be fixed.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>