QOS ellision for wavefront sink, CKMS resets
This commit introduces QOS ellison back into the wavefront sink.
For metrics which maintain their timestamp this is done by seeking
through the bucket's vec buffer and discarding points that are too
close together in time. For CKMS structures--which are now reset
on every successful flush--we keep a notion of at what times the
structures were snapshotted in the wavefront structure.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>