Avoid reading 1MB every mpsc iteration
Guess what! read_exact has no notion of slices, counter to what
I was clearly expecting when I wrote mpsc. The reason for our
falling behind was _something_ in mpsc and it turns out it was
forcing mpsc to read 1MB each iteration, throwing away all but a
tiny bit of that read.

Cool.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>