Throw a BufReader around the Receiver fp
Previously this wasn't possible to do as the Receiver was meant
to be shared between threads. However! Now that the Receiver is
pegged to a single thread we can BufRead the underlying fp. This
doesn't save _too much_ time but does cut down on our syscalls.

Every little bit helps.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>