QOS elision (#102)
* Improve testing for buckets

This commit ramps up the quickcheck testing we're doing of the
bucket module. The ultimate aim here is to get at QOS violations
but first we need to model the behaviour of the system.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* No more extra snapshot thread

All metrics carry in themselves a timestamp. With this, it's
possible for the aggregation module--buckets--to be aware of
if a metric is inside or outside the bounds of the metric's
QOS. Because this implementation was complicated we skipped
it previously and relied on a snapshotting thread--which
fired on one second intervals--to accomplish kinda sorta the
same thing.

This commit removes the snapshotting thread, the snapshot
event and pushes the concept of snapshotting into the metric
aggregtor directly. What we don't do is obey QOS notions yet
which will be fixed.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Buckets benchmarks, qos removal from buckets

The concept of QOS is now removed from the buckets aggregation
as this, I believe, is rightly something a sink should be
aware of. That is, a bucket will bin on one-second intervals
and sinks may choose to QOS based on that at their convenience.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Test bucket reset, correct gauge reset

This commit drops all but the last point in the gauge
vec and tests the good working order of reset via a new
QC test. Previously we would steadily accumulate gauge
points through time.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* QOS ellision for wavefront sink, CKMS resets

This commit introduces QOS ellison back into the wavefront sink.
For metrics which maintain their timestamp this is done by seeking
through the bucket's vec buffer and discarding points that are too
close together in time. For CKMS structures--which are now reset
on every successful flush--we keep a notion of at what times the
structures were snapshotted in the wavefront structure.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* cargo fmt run

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Only reset histograms / timers when appropriate

Previously we would reset histograms / timers every flush interval
even if we had not yet reached their QOS period. This would result
in data loss. This does not affect the other metric types which have
an associated QOS because they are obligated to ship at least one
point per flush.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Reset console aggrs

This commit resets the console aggregation buffers on each
flush.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add trace info into server

It would be useful to see statsd packets as trace information
in a like fashion to the way we see graphite packets. This
commit allows that.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Correct gauge / counter regression

Turns out, we were eliding too many points. The last sample
counter is meant to be per-metric but its placement outside of
the metric loop made it global. Oops.

This is correct and tests are in place to verify the correction.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove surprise in mpsc with regard to log file size

It turns out that if you use the default channel_0 to create your
mpsc channels the byte value you put in will be multiplied by 100.
The intention--undocumented--was that this should be 10. So!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Use native CPU instructions, add benchmark

I'm trying to track down why it takes 50M nanoseconds to
do the send/recv benchmark for mpsc. This is one of the
more important bits in cernan. I don't fully know what
the deal is at present.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Produce a cernan-unstable binary

This commit modifies the Jenksinsfile to produce a cernan-unstable
artifact in s3, which will make our lives a little bit easier.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Bump channel queue file to 100MB

In practice we chew through 10MB files very quickly on mid-load
systems. This might properly be a config option.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add null sink

We have a fun issue where it's not clear on a production system
if the wavefront sink is falling behind because _it_ is slow or
if it is falling behind because mpsc is slow. So! Here's a sink
that immediately discards the metrics it receives.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Avoid reading 1MB every mpsc iteration

Guess what! read_exact has no notion of slices, counter to what
I was clearly expecting when I wrote mpsc. The reason for our
falling behind was _something_ in mpsc and it turns out it was
forcing mpsc to read 1MB each iteration, throwing away all but a
tiny bit of that read.

Cool.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add an additional mpsc benchmark

This benchmark should always return in-line with the send/recv
benchmark which already exists. _Should_.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Throw a BufReader around the Receiver fp

Previously this wasn't possible to do as the Receiver was meant
to be shared between threads. However! Now that the Receiver is
pegged to a single thread we can BufRead the underlying fp. This
doesn't save _too much_ time but does cut down on our syscalls.

Every little bit helps.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>