Introduce the Kinesis Firehose sink
We'd like to use Kinesis Firehose to plonk data into S3 and
ElasticSearch conveniently. This commit makes use of the
rusoto library to do the plonking. The main bulk of things
is the loop in firehose flush_0 which packages log lines
as json and flips them out over the wire. This loop works
within the limits of Firehose, or attempts to.

Note that we're now making the distinction between a 'Metric',
considered to be a single point-in-time, and a LogLine, considered
to be an unparsed but potentially interesting line of
information-in-time.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>