MPSC improvements (#104)
* Make mpsc generic

This commit makes the mpsc structure generic with regard to serialized
type. The fun bit here is the use of PhantomData to keep type-safety
even though the Sender and Receiver never actually store T in their
own memory.

Weeeee type systems!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Increase mpsc tests, correct concurrent flaws

In extreme situations--where both the sender and receiver are
aggressively fast and queue file byte size is very small--there
were concurrency bugs evident in mpsc. These are now corrected.

In particular:

   * The state-machine inside of the sender was overly
     complicated and has been reduced with a compare-and-swap
     based implementation which is more correct and faster.
   * The receiver could abandon data when switching over queue
     files if it was hot on the heels of the senders, corrected
     with a retry loop.
   * Searches of the filesystem are now down to interpret file
     paths as integers, corrected a seq_num lookup bug in both
     Sender and Receiver when there are many files on-disk.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Incorporate John Koenig's suggestions

This commit fixes a one-byte bug and adds logs around the potentially
infinite but in-practice bounded loops in mpsc.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>