Reduce CPU use of mpsc, corruption on certain file-systems
This commit reduces the CPU load of mpsc by changing how
Receiver::next works. Now, None will be returned if data is not
immediately available on-disk, complicating streaming data
from this Iterator but allowing the client to choose how much
CPU to dedicate to the task. The Sender and Receiver coordinate
via an AtomicUsize to track how many payloads are available,
allowing for efficient streams of Somes vs. before.

Additionally, data corruption was detected on ext4 filesystems on
account of

    https://stackoverflow.com/questions/32851672/is-overwriting-a-small-file-atomic-on-ext4

The fs2 library is used to get access to advisory locking. The
Sender takes a shared lock on the file handler while the Reader
demands an exclusive one. There is a potential for starvation
here but, in practice and under test, this does not seem to be an
issue.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>