Remove fs2, advisory locks
We observed under test on OSX a deadlock owing to the file locks.
Probably. We didn't dig in because we're bums. Instead, wanting
to preserve behaviour on systems which don't have convenient
advisory locks, we've elected to use a userland mutex between all
senders and receivers.

The critical sections are relatively small, though there is some
thought that, if we're gonna lock things we might choose to avoid
some of the other busy-loop synchronizations. Dunno!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>