Modularize mpsc, go all-in on mutexes
In the spirit of Joe's "Make it right, then make it fast" and with
increased knowledge of how wacky filesystems are in practice--did
you know that technically ext4 violates POSIX rules on atomic
reads and writes--we're going all-in on mutual exclusion.

This commit also splits the sender and receiver into distinct
sub-modules. This is done because I found it increasingly difficult
to find my place in their files, what with all the documentation.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>