Rough cut of cernan to cernan communication
This is a demonstration of the rough fashion in which cernan to
cernan communication could be made to work. The wire-level protocol
is cribbed heavily from mpsc and if you try and start this thing up
it will crash for reasons that need investigation. Basic idea, like
I say.

Related to #110

Signed-off-by: Brian L. Troutwine <blt@postmates.com>