Reduce CPU use of mpsc, corruption on certain file-systems (#108)
* Reduce CPU use of mpsc, corruption on certain file-systems

This commit reduces the CPU load of mpsc by changing how
Receiver::next works. Now, None will be returned if data is not
immediately available on-disk, complicating streaming data
from this Iterator but allowing the client to choose how much
CPU to dedicate to the task. The Sender and Receiver coordinate
via an AtomicUsize to track how many payloads are available,
allowing for efficient streams of Somes vs. before.

Additionally, data corruption was detected on ext4 filesystems on
account of

    https://stackoverflow.com/questions/32851672/is-overwriting-a-small-file-atomic-on-ext4

The fs2 library is used to get access to advisory locking. The
Sender takes a shared lock on the file handler while the Reader
demands an exclusive one. There is a potential for starvation
here but, in practice and under test, this does not seem to be an
issue.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove fs2, advisory locks

We observed under test on OSX a deadlock owing to the file locks.
Probably. We didn't dig in because we're bums. Instead, wanting
to preserve behaviour on systems which don't have convenient
advisory locks, we've elected to use a userland mutex between all
senders and receivers.

The critical sections are relatively small, though there is some
thought that, if we're gonna lock things we might choose to avoid
some of the other busy-loop synchronizations. Dunno!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Improve benchmarks for mpsc

This commit improves the benchmarks by reducing the amount of
non-test data that must be created per iteration.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Panic if attempts > 60_000

We saw a run on Jenkins where a test looped and failed because
attempts += 1 eventually hit its max size. This means that we
had outstanding writes to read but Receiver could not find them.

It's our hope to trigger this condition and get more information.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Modularize mpsc, go all-in on mutexes

In the spirit of Joe's "Make it right, then make it fast" and with
increased knowledge of how wacky filesystems are in practice--did
you know that technically ext4 violates POSIX rules on atomic
reads and writes--we're going all-in on mutual exclusion.

This commit also splits the sender and receiver into distinct
sub-modules. This is done because I found it increasingly difficult
to find my place in their files, what with all the documentation.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add variable length write QC test

This commit adds a new test qc_concurrent_snd_and_rcv_round_trip
patterned off the existing non-qc test which writes randomized,
variable length data to disk.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update cernan mpsc docs

This commit updates the mpsc docs to discuss the results of
experimenting with different filesystems under load. The Sender::new
and Receiver::new functions are public and documented as "PRIVATE"
at the moment because I don't know how to clean them up otherwise.

Someday!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>