Wavefront flush delays (#111)
* Delay wavefront sink in the event of flush failures

Prior to this commit the wavefront sink would continue to pull
and buffer points in its buckets if it were partitioned away
from its report-proxy. This was not so bad when the bucket
concept was Atom -> f64 but is problematic now that it is
Atom -> Vec<Metric>.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Loop only in flush_1, revert to default Sink run for Wavefront

As John pointed out, the previous implementation had the nifty
side-effect of _never_ re-establishing our connection to Wavefront,
leaving us high and dry. This now blocks in the flush attempt,
resulting in the same system behaviour but without the added
brokeness.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>