Update to 0.4.0
This release incorporates work to move away from std's in-memory channel,
freeing up cernan to emit toward slow sinks--firehose--and ingest more
data for analysis that will fit into memory. Performance remains on par
with the previous cernan series and has been improved in regards to
statsd/graphite parse speed.

cernan is now well arranged to allow in-flight analysis and manipulation
of data, as well as forward off-system to itself for aggregation should
the present system be particularly small but heavily telemetered.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>