Allow statsd / graphite to be optional
Previously cernan _had_ to start a statsd and graphite interface.
This is awkward for some folks. It's now possible to turn off one
or both interfaces, though they remain on by default.

This commit changes the configuration file format while keeping
backward compatability. It's not possible to disable an interface
without using the new configuration style.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>