Get cernan to cernan communication going
This commit makes it possible for us to properly ship events across
the wire to a remote cernan. The protocol is a length prefixed frame
that has been zlib compressed. Bincode, our serialization format, is
very fast to encode / decode but not especially small. It does compress
quite well, going from 50k bytes to 300 bytes in test runs.

The wire-level protocol is similar to the mpsc file format and maybe
we ought to experiment with compression there, too.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>