Correct QOS sampling for sparse timeseries
At the moment the way we compute QOS elision for sparse timeseries
is wrong. We have assumed in our algorithm that all timeseries will
be dense, allowing us to elide purely based on time. However,
consider a QOS set to 15 and points that arrive at the 15th, 30th,
45th and 60th second. These points will be elided, causing the
timeseries to not be reported.

This commit addresses that problem. Given a flush interval F and a QOS Q the
maximum number of points that will be emitted for the interval F will
be `(1/Q)*F`. In the above, let `Q = 60`. We expect the number
of points to be reported as `(1/15)*60 = 4` and this commit makes
that happen.

Consider as well the case where points are emitted with the above
givens at the following times:

  * T15
  * T30
  * T35
  * T45
  * T60

There are five points: one must be elided. To avoid introducing any
systematic artifacts, we _randomly_ choose the maximum points to
be emitted for any given interval F. That is, this commit will pick
four of the five points randomly and ship those, eliding one.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>