Cernan to Cernan (#119)
* Rough cut of cernan to cernan communication

This is a demonstration of the rough fashion in which cernan to
cernan communication could be made to work. The wire-level protocol
is cribbed heavily from mpsc and if you try and start this thing up
it will crash for reasons that need investigation. Basic idea, like
I say.

Related to #110

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Get cernan to cernan communication going

This commit makes it possible for us to properly ship events across
the wire to a remote cernan. The protocol is a length prefixed frame
that has been zlib compressed. Bincode, our serialization format, is
very fast to encode / decode but not especially small. It does compress
quite well, going from 50k bytes to 300 bytes in test runs.

The wire-level protocol is similar to the mpsc file format and maybe
we ought to experiment with compression there, too.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add configurable ip to CRD, documentation

This commit resolves the two things I realized were deficient about
the introduction of CRD. In particular, the receiver now has a
configurable host and the README has been updated to reflect this
new capability.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Change sinks/cernan.rs -> sinks/crd_receiver.rs

The module was originally named before John and I came up with the
name for the subsystem as a whole. Leaving src/cernan.rs would have
been confusing.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remvoe println

This commit removes a println left behind while I was developing.
There's a need for more effective logging in cernan and here's the
ocassional goof to prove it.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* crd -> federation

This commit changes the nomenclature of the CRD to Federation,
at the suggestion of Tom. It is a clearer name. I stuck to
receiver / transmitter because I found it difficult to understand
documentation which discussed the federation_sink being a source
for other sinks.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>