File Source Cleanup (#124)
* Add tests around multi-file sources, adjust firehose

DAK had a crash when working with multi-file sources. I didn't get
the full config but went ahead and added more tests. I was not able
to reproduce.

This commit also adjust the format that timestamp goes out in, being
ISO8601. This should make Elasticsearch happier.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Adjust time format for kibana

We've got a peculiar deal where Elasticsearch can parse the timestamp
that cernan emits but kibana cannot. Cool beans.

DAK hopes that if we adjust the timestamp _just so_ we'll be good to
go.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>