Confirm sampling at the wavefront level
I noticed something weird in the wavefront reporting. When flush interval
is set to 60s and the qos for a timer is 15 there will be _five_ points per
one minute interval. Now, it's not the case that cernan starts its timer
on the :00 of each minute so this is--give the confirmation tests--either
an offset artifact or something goofy in the daisy chain of datastores
through to the visualization.

Dunno!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>