Correct QOS sampling for sparse timeseries (#126)
* Correct QOS sampling for sparse timeseries

At the moment the way we compute QOS elision for sparse timeseries
is wrong. We have assumed in our algorithm that all timeseries will
be dense, allowing us to elide purely based on time. However,
consider a QOS set to 15 and points that arrive at the 15th, 30th,
45th and 60th second. These points will be elided, causing the
timeseries to not be reported.

This commit addresses that problem. Given a flush interval F and a QOS Q the
maximum number of points that will be emitted for the interval F will
be `(1/Q)*F`. In the above, let `Q = 60`. We expect the number
of points to be reported as `(1/15)*60 = 4` and this commit makes
that happen.

Consider as well the case where points are emitted with the above
givens at the following times:

  * T15
  * T30
  * T35
  * T45
  * T60

There are five points: one must be elided. To avoid introducing any
systematic artifacts, we _randomly_ choose the maximum points to
be emitted for any given interval F. That is, this commit will pick
four of the five points randomly and ship those, eliding one.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove left-over prints

This commit removes a small contingent of println statements that
were left unintentionally. Womp womp

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Instrument the wavefront QOS

Each time we emit to wavefront it'd be very useful to know what
our QOS values are. This is especially important for setting up
a controlled experiment into sampling.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Quickcheck the sampling

After an intitial experiment it was clear there was some odd
behaviour coming out of the sampling system. This commit introduces
a quickcheck test to assert that

  1. all sampled vectors are of the same size as a sampled
     vector, once you know the total intervals in said vector

  2. none of the sampled interval runs have more than the allowed
     sample points.

This was fairly tricky to do and I've put many example unit tests
in place as well.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Confirm sampling at the wavefront level

I noticed something weird in the wavefront reporting. When flush interval
is set to 60s and the qos for a timer is 15 there will be _five_ points per
one minute interval. Now, it's not the case that cernan starts its timer
on the :00 of each minute so this is--give the confirmation tests--either
an offset artifact or something goofy in the daisy chain of datastores
through to the visualization.

Dunno!

Signed-off-by: Brian L. Troutwine <blt@postmates.com>